+++ KompresjaWAV/src/test/java/pg/eti/ksd/kompresjawav/engine/CompressedPacketImplTest.java
@@ -37,14 +37,14 @@ public class CompressedPacketImplTest {
 
     @Test
     public void test_computeQuantizationLevels_generatesCorrectNumberOfLevels() {
-        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Double[]{1.0}), Arrays.asList(new Double[]{1.0, 2.0, 3.0}), Arrays.asList(new Sample[]{}));
+        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Float[]{1.0f}), Arrays.asList(new Double[]{1.0, 2.0, 3.0}), Arrays.asList(new Sample[]{}));
         List<Double> actual = sut.computeQuantizationLevels(CompressedPacketImpl.QUANTIZATION_LEVELS);
         Assert.assertEquals(CompressedPacketImpl.QUANTIZATION_LEVELS, actual.size());
     }
 
     @Test
     public void test_computeQuantizationLevels_generatesCorrect4Levels() {
-        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Double[]{1.0}), Arrays.asList(new Double[]{1.0, 2.0}), Arrays.asList(new Sample[]{}));
+        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Float[]{1.0f}), Arrays.asList(new Double[]{1.0, 2.0}), Arrays.asList(new Sample[]{}));
         List<Double> actual = sut.computeQuantizationLevels(4);
         Assert.assertEquals(-3 / 2.0, actual.get(0));
         Assert.assertEquals(-1 / 2.0, actual.get(1));
@@ -59,7 +59,7 @@ public class CompressedPacketImplTest {
             errors.add(1.0);
         }
 
-        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Double[]{1.0}), errors, Arrays.asList(new Sample[]{}));
+        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Float[]{1.0f}), errors, Arrays.asList(new Sample[]{}));
         List<Long> compressed = sut.compressErrors(errors);
         Assert.assertEquals(1, compressed.size());
     }
@@ -71,7 +71,7 @@ public class CompressedPacketImplTest {
             errors.add(0.0);
         }
 
-        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Double[]{1.0}), errors, Arrays.asList(new Sample[]{}));
+        CompressedPacketImpl sut = new CompressedPacketImpl(Arrays.asList(new Float[]{1.0f}), errors, Arrays.asList(new Sample[]{}));
         List<Long> compressed = sut.compressErrors(errors);
         Assert.assertEquals(0, (long) compressed.get(0));
     }
